Fix sass webpack setup | refactor button.js for sass | clean up global scope for button.js
