--- src/components/button/Button.css
@@ -1,5 +1,0 @@
-.button {
-    padding: 10px 16px;
-    cursor: pointer;
-    text-transform: capitalize;
-}
