+++ src/components/button/Button.scss
@@ -1,0 +1,7 @@
+button {
+    &.button {
+        padding: 10px 16px;
+        cursor: pointer;
+        text-transform: capitalize;
+    }
+}
