+++ src/components/rate/Rate.scss
