+++ src/pages/rates/Rates.js
@@ -1,6 +1,6 @@
 import RatesStore from "../../store/Rates";
 import Rate from "../../components/rate/Rate";
-import RateStyle from "./Rates.css" assert {type: 'css'};
+import RateStyle from "./Rates.scss";
 
 const ERROR_TEMPLATE = '<p class="message-error">Something went wrong</p>' //Hard coded error message, not good for production application
