+++ src/pages/rates/Rates.scss
